move remaining of hey logic to peermgr
